[HOTSPOT-174] test: 테스트 코드 도메인 별 분리
